Move SilentExceptionWrapper from testing helper to primary option
